+++ src/corelib/text/qbytearray.cpp
@@ -636,7 +636,7 @@ QByteArray qUncompress(const uchar* data, qsizetype nbytes)
     size_t expectedSize = size_t((data[0] << 24) | (data[1] << 16) |
                                  (data[2] <<  8) | (data[3]      ));
     size_t len = qMax(expectedSize, 1ul);
-    const ulong maxPossibleSize = MaxAllocSize - sizeof(QByteArray::Data);
+    const size_t maxPossibleSize = MaxAllocSize - sizeof(QByteArray::Data);
     if (Q_UNLIKELY(len >= maxPossibleSize)) {
         // QByteArray does not support that huge size anyway.
         return invalidCompressedData();
